+++ README.md
@@ -1,2 +1,12 @@
 # retail-url-scraper
 A web scraper that scrapes retail websites and returns URLs for all listed products. Runs on Express and Node
+
+To Install and run:
+
+1. npm install
+
+2. DEBUG=site-urls:* npm start
+
+
+node v10.16.3
+npm v6.9.0
