+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "node ./bin/www"
+    "start": "DEBUG=gzipservice* node ./app/start-server.js"
   },
   "dependencies": {
     "cookie-parser": "~1.4.4",
